+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/Synchronizer.kt
@@ -266,6 +266,9 @@ interface Synchronizer {
         memo: String = ZcashSdk.DEFAULT_SHIELD_FUNDS_MEMO_PREFIX
     ): Long
 
+    // TODO [#1534]: Add RustLayerException.ValidateAddressException
+    // TODO [#1534]: https://github.com/Electric-Coin-Company/zcash-android-wallet-sdk/issues/1534
+
     /**
      * Returns true when the given address is a valid z-addr. Invalid addresses will throw an
      * exception. See valid z-addresses characteristics in related ZIP.
@@ -303,6 +306,21 @@ interface Synchronizer {
      * @throws RuntimeException when the address is invalid.
      */
     suspend fun isValidUnifiedAddr(address: String): Boolean
+
+    /**
+     * Returns true when the given address is a valid ZIP 320 TEX address.
+     *
+     * This method is intended for type checking (e.g. form validation). Invalid
+     * addresses will throw an exception. See valid t-addresses characteristics
+     * in the related ZIP.
+     *
+     * @param address the address to validate.
+     *
+     * @return true when the given address is a valid ZIP 320 TEX address.
+     *
+     * @throws RuntimeException when the address is invalid.
+     */
+    suspend fun isValidTexAddr(address: String): Boolean
 
     /**
      * Validate whether the server and this SDK share the same consensus branch. This is